# backend/internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EpicPanel API",
    "version": "1.0.0",
    "description": "Linux-first hosting control plane API. Session cookie or Bearer session token for users; Bearer epk_ tokens for automation (scoped)."
  },
  "servers": [{"url": "/"}],
  "components": {
    "securitySchemes": {
      "bearer": {"type": "http", "scheme": "bearer"}
    },
...
  },
  "paths": {
...
    "/v1/organizations/{org_id}/databases/{db_id}/pma-sso": {"get": {"summary": "One-time phpMyAdmin/Adminer SSO link (developer+, audited)", "security": [{"bearer": []}], "responses": {"200": {"description": "url"}, "409": {"description": "not ready"}}}},
...
    "/v1/organizations/{org_id}/databases": {
      "get": {"summary": "List databases (billing+)", "security": [{"bearer": []}], "responses": {"200": {"description": "ok"}}},
      "post": {"summary": "Create database {name, server_id, engine, website_id?} (developer+) -> 202", "security": [{"bearer": []}], "responses": {"202": {"description": "creating"}}}
    },
    "/v1/organizations/{org_id}/databases/{db_id}/credentials": {"get": {"summary": "Reveal credentials (developer+, audited)", "security": [{"bearer": []}], "responses": {"200": {"description": "ok"}, "409": {"description": "not available"}}}},
...
    "/v1/organizations/{org_id}/api-tokens": {
      "get": {"summary": "List API tokens (admin+)", "security": [{"bearer": []}], "responses": {"200": {"description": "ok"}}},
      "post": {"summary": "Create API token {name, scopes, expires_in_days?} (admin+) — raw shown once", "security": [{"bearer": []}], "responses": {"201": {"description": "created"}}}
    },
    "/v1/organizations/{org_id}/api-tokens/{token_id}": {"delete": {"summary": "Revoke API token (admin+)", "security": [{"bearer": []}], "responses": {"204": {"description": "done"}}}},
...
  }
}